+++ ProjApp.Database/Entities/InitialVerificationFailed.cs
@@ -13,7 +13,7 @@ public class InitialVerificationFailed : DatabaseEntity, IInitialVerification
 
     // Optional
     public string? VerificationTypeNum { get; set; }
-    public uint? OwnerInn { get; set; }
+    public ulong? OwnerInn { get; set; }
     public string? Worker { get; set; }
     public DeviceLocation? Location { get; set; }
     public string? AdditionalInfo { get; set; }
