+++ ProjApp.Database/Entities/Owner.cs
@@ -3,5 +3,5 @@ namespace ProjApp.Database.Entities;
 public class Owner : DatabaseEntity
 {
     public required string Name { get; set; }
-    public required uint INN { get; set; }
+    public required ulong INN { get; set; }
 }
